+++ include/tmc/detail/thread_layout.ipp
@@ -43,7 +43,7 @@ static void recursively_subdivide(std::vector<SubdivideNode>& Results) {
 // path structure except for that branch. Keeping the original path structure
 // produces a more balanced work stealing network / matrix.
 static void enumerate_paths(
-  const size_t Path, uint64_t DepthBit,
+  const size_t Path, size_t DepthBit,
   const std::vector<SubdivideNode>& PathTree, const size_t NodeIdx,
   std::vector<size_t>& Results
 ) {
@@ -108,7 +108,7 @@ get_group_iteration_order(size_t GroupCount, size_t StartGroup) {
   // the high bits are padded with 0s (0b0...00010110), which solves the
   // case of a start node on the short (high) side of the tree
 
-  uint64_t startPath = 0;
+  size_t startPath = 0;
   SubdivideNode node = pathTree[0];
   size_t depth = 0;
   {
@@ -253,7 +253,7 @@ void bind_thread(hwloc_topology_t Topology, hwloc_cpuset_t SharedCores) {
     std::vector<unsigned long> bitmapUlongs;
     bitmapUlongs.resize(bitmapSize);
     hwloc_bitmap_to_ulongs(SharedCores, bitmapSize, bitmapUlongs.data());
-    std::vector<uint64_t> bitmaps;
+    std::vector<size_t> bitmaps;
     if constexpr (sizeof(unsigned long) == 8) {
       bitmaps.resize(bitmapUlongs.size());
       for (size_t b = 0; b < bitmapUlongs.size(); ++b) {
@@ -271,7 +271,7 @@ void bind_thread(hwloc_topology_t Topology, hwloc_cpuset_t SharedCores) {
         if (b >= bitmapUlongs.size()) {
           break;
         }
-        bitmaps.back() |= ((static_cast<uint64_t>(bitmapUlongs[b])) << 32);
+        bitmaps.back() |= ((static_cast<size_t>(bitmapUlongs[b])) << 32);
         ++b;
       }
     }
